Luxuria Domus Rental System (Web Backed Up)
This is the code for the engine that runs the Luxuria Domus Rental
System for OpenSim Grids.
LD Rental Board is the engine for the board people pay to rent units
Luxuria Domus Rental Server is the engine for the Main Server that
tracks everything and backs up data Off-World.
Sign Changer Script is the engine that can be placed into several prims
to make a unit available board.